+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Todos los cambios notables de este proyecto se documentan en este archivo.
 
-El formato está basado en [Keep a Changelog](https://keepachangelog.com/es/1.0.0/)
+El formato está basado en [Keep a Changelog](https://keepachangelog.com/es-ES/1.1.0/)
 y este proyecto adhiere a [Conventional Commits](https://www.conventionalcommits.org/).
 
 ---